+++ extensions/flac/src/androidTest/java/com/google/android/exoplayer2/ext/flac/FlacPlaybackTest.java
@@ -76,12 +76,10 @@ public class FlacPlaybackTest extends InstrumentationTestCase {
       DefaultTrackSelector trackSelector = new DefaultTrackSelector();
       player = ExoPlayerFactory.newInstance(new Renderer[] {audioRenderer}, trackSelector);
       player.addListener(this);
-      ExtractorMediaSource mediaSource = new ExtractorMediaSource(
-          uri,
-          new DefaultDataSourceFactory(context, "ExoPlayerExtFlacTest"),
-          MatroskaExtractor.FACTORY,
-          null,
-          null);
+      ExtractorMediaSource mediaSource = new ExtractorMediaSource.Builder(
+          uri, new DefaultDataSourceFactory(context, "ExoPlayerExtFlacTest"))
+          .setExtractorsFactory(MatroskaExtractor.FACTORY)
+          .build();
       player.prepare(mediaSource);
       player.setPlayWhenReady(true);
       Looper.loop();